requirements.txt
absl-py==0.12.0
academictorrents==2.3.3
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
Automat==0.6.0
backcall==0.1.0
bencode.py==2.0.0
bitstring==3.1.5
bleach==3.1.5
boto3==1.17.93
botocore==1.20.93
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
chart-studio==1.0.0
cleverhans @ git+https://github.com/tensorflow/cleverhans.git@1115738a3f31368d73898c5bdd561a85a7c1c741
click==8.0.1
clip @ git+https://github.com/openai/CLIP.git@cfcffb90e69f37bf2ff1e988237a0fbe41f33c04
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
cryptography==2.1.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distro-info===0.18ubuntu0.18.04.1
dm-tree==0.1.6
easydict==1.9
edward2 @ git+https://github.com/google/edward2.git@d2571a25bd4ed4a4575f4f64f5048b5e1a8bf233
efficientnet-pytorch==0.7.1
entrypoints==0.3
fastdtw==0.3.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
ftfy==6.0.3
future==0.16.0
gast==0.4.0
glob3==0.0.1
google-api-core==1.31.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
graphviz==0.5.2
grpcio==1.39.0
h5py==3.1.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
imageio==2.5.0
importlib-metadata==4.0.1
importlib-resources==5.1.2
incremental==16.10.1
ipykernel==5.2.1
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.13.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.2
jupyterlab-server==1.1.3
Keras-Applications==1.0.8
keras-nightly==2.7.0.dev2021072600
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
libclang==11.1.0
llvmlite==0.33.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.1.1
mistune==0.8.4
ml-collections==0.1.0
mnist==0.2.2
mplcursors==0.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
networkx==2.3
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.3
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
pluggy==0.6.0
pretrainedmodels==0.7.4
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.17.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.1
Pygments==2.6.1
pygobject==3.26.1
pyOpenSSL==17.5.0
pyparsing==2.4.7
PyPubSub==3.3.0
pyrsistent==0.16.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.1
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
retrying==1.3.3
rfpimp==1.3.4
robustness-metrics @ git+https://github.com/google-research/robustness_metrics.git@c08df4a86fee978d38ef6eba8a41b9e483000c5b
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.96
service-identity==16.0.0
six==1.15.0
sklearn==0.0
sos==4.1
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.37.6
ssh-import-id==5.7
statsmodels==0.12.2
systemd-python==234
tabulate==0.8.9
tb-nightly==2.6.0a20210720
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.1.0
tensorflow-probability==0.12.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.7.0.dev2021072101
tf-nightly==2.7.0.dev20210721
tfa-nightly==0.14.0.dev20210710011535
tfds-nightly==4.3.0.dev202107260106
tfp-nightly==0.14.0.dev20210726
timm==0.4.9
torch==1.7.1
torchvision==0.8.2
tornado==6.0.4
tox==2.5.0
tqdm==4.60.0
traitlets==4.3.3
Twisted==17.9.0
typeguard==2.12.1
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uncertainty-baselines==0.0.7
urllib3==1.26.4
virtualenv==15.1.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.0.0
wrapt==1.12.1
xdot==0.9
xkit==0.0.0
zipp==3.4.1
zope.interface==4.3.2
